+++ src/lib/state/eventStore.ts
@@ -36,13 +36,16 @@ class EventStore {
    * Removes the argued event by ID if all reminders have occurred or
    * scheduled date is passed
    * @param {string} id
+   * @returns {boolean}
    * @memberof EventStore
    */
-  removeIfOld(id: string) {
+  removeIfOld(id: string): boolean {
     const e: CalendarEvent | undefined = this._events.get(id)
     if (e !== undefined && [...e.reminders.values()].every(v => v)) {
       this._events.delete(id)
+      return true
     }
+    return false
   }
 
   /**